Create TSP
TSP and TSP_berlin52: using the Berlin52 dataset, the most famous of the problem. The Berlin52 is a list of 52 locations and their coordinates in a Cartesian coordinate system overlaid on the city of Berlin. In this example we will build an initial solution using the Nearest algorithm Neighbor, where in each location we simply select the nearest point as the following on the route [8]. To improve this initial solution, we will then implement the 2-opt algorithm