Add files via upload
Controlo Orçamental: to make the budget control we have some current challenges that happen in the business world. First to be able to relate tables to each other from the date we have to create a Data column. So, here I created a column with the desired aggregation level, in this case the month for each year to be able to relate the tables. In addition, the budget is designed by the month and we want to disaggregate at the level of the day.